+++ login.php
@@ -617,7 +617,7 @@ class LoginPlugin extends Plugin
             $this->grav['assets']->add('plugin://login/css/login.css');
         }
 
-        $task = $this->grav['uri']->param('task');
+        $task = $this->grav['uri']->param('task') ?: isset($_POST['task']) ? $_POST['task'] : '';
         $task = substr($task, strlen('login.'));
         if ($task === 'reset') {
             $username = $this->grav['uri']->param('user');
@@ -627,7 +627,8 @@ class LoginPlugin extends Plugin
                 $twig->twig_vars['username'] = $username;
                 $twig->twig_vars['token'] = $token;
             }
-
+        } elseif ($task === 'login') {
+            $twig->twig_vars['username'] = isset($_POST['username']) ? $_POST['username'] : '';
         }
     }